Add server requirements (#67)
* Add server requirements

Adding requirements for the server.
This is necessary, if you want to use PUSH or TiQR token.
closes #60

* Added message if server version too low

If the administrator checks the service credentials, he will now get an
information, if the privacyIDEA server version is recommended.

* Changed string to array

Now we can enter multiple recommended versions.
If e.g. 3.x and 2.23.5 are supported, we could enter
"3.", "2.23.5".